# api/openapi.yaml
openapi: 3.1.0
info:
  title: Santé Rurale Mali API
  description: |
    API REST pour la gestion des dossiers patients en zones rurales du Mali.

    ## Caractéristiques
    - Offline-first avec synchronisation opportuniste
    - Authentification JWT (RS256)
    - Support FHIR R4 et exports DHIS2
    - Gestion multi-sites avec RBAC

    ## Sécurité
    - TLS 1.3 requis en production
    - JWT access token (15 min) + refresh token (7 jours)
    - Rate limiting: 100 req/min par IP, 1000 req/heure par user

  version: 1.0.0
  contact:
    name: Support Technique
    email: support@sante-rurale.ml
servers:
  - url: https://api.sante-rurale.ml/v1
    description: Production
  - url: https://staging-api.sante-rurale.ml/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Développement local
...
security:
  - BearerAuth: []
paths:
...
  /patients:
    get:
      tags: [patients]
      summary: Liste des patients
      description: Retourne les patients du site de l'utilisateur (ou tous si admin)
      parameters:
        - name: search
          in: query
          description: Recherche par nom/prénom/téléphone
          schema:
            type: string
            example: Diara
        - name: village
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste des patients
          headers:
            X-Total-Count:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [patients]
      summary: Créer un patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Patient créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /patients/{id}:
    get:
      tags: [patients]
      summary: Détails d'un patient
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Détails du patient
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDetails'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [patients]
      summary: Modifier un patient
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Patient modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtenu via /auth/login.
        Format: `Authorization: Bearer <token>`
  parameters:
    PatientId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Curseur de pagination
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Nombre d'éléments par page
      schema:
        type: integer
        default: 50
        maximum: 200
    IfMatch:
      name: If-Match
      in: header
      description: ETag pour la gestion de concurrence optimiste
      schema:
        type: string
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nom:
          type: string
          example: Sanogo
        prenom:
          type: string
          example: Diara
        sexe:
          type: string
          enum: [M, F]
          example: F
        annee_naissance:
          type: integer
          example: 1978
        age:
          type: integer
          example: 46
          description: Calculé automatiquement
        telephone:
          type: string
          example: "+223 70 12 34 56"
        village:
          type: string
          example: Konobougou
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          description: Version pour gestion de conflits
    PatientCreate:
      type: object
      required: [nom, sexe]
      properties:
        nom:
          type: string
          minLength: 2
        prenom:
          type: string
        sexe:
          type: string
          enum: [M, F]
        annee_naissance:
          type: integer
          minimum: 1900
          maximum: 2025
        telephone:
          type: string
        village:
          type: string
    PatientUpdate:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        village:
          type: string
    PatientDetails:
      allOf:
        - $ref: '#/components/schemas/Patient'
        - type: object
          properties:
            recent_encounters:
              type: array
              items:
                $ref: '#/components/schemas/Encounter'
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
    Encounter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        patient:
          $ref: '#/components/schemas/Patient'
        site_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user:
          type: object
          properties:
            nom:
              type: string
        date:
          type: string
          format: date
        motif:
          type: string
        temperature:
          type: number
          format: float
          example: 36.8
        pouls:
          type: integer
          example: 78
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        encounter_id:
          type: string
          format: uuid
        filename:
          type: string
        s3_key:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          description: Curseur actuel
        next_cursor:
          type: string
          description: Curseur pour la page suivante
        has_more:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
          example: Le champ 'nom' est requis
        details:
          type: object
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflit (ex. doublon, violation contrainte)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PreconditionFailed:
      description: Échec de la condition préalable (ETag mismatch)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...